Apply gamma correction to UI LEDs.
